+++ packages/mdx-content/src/server.ts
@@ -1,10 +1,9 @@
 export { ButtonMdx } from "./components/ButtonMdx";
-export { CardMdx } from "./components/CardMdx";
 export { CardsMdx } from "./components/CardsMdx";
 export { CodeBlock } from "./components/CodeBlockMdx";
-export { FileMdx, FilesMdx, FolderMdx } from "./components/FileTreeMdx";
 export { KbdMdx } from "./components/KbdMdx";
 export { LinkMdx } from "./components/LinkMdx";
+export { NoteMdx } from "./components/NoteMdx";
 export { ChangesMdx, ReleaseMdx } from "./components/ReleaseMdx";
 export { StepMdx, StepsMdx } from "./components/StepperMdx";
 export {